[FE-127] Moved the Collateral componenet to the ContentUser component.
