+++ src/App.jsx
@@ -15,7 +15,6 @@ import Borrow from './components/Borrow/Borrow';
 import RepayAll from './components/Repay/RepayAll';
 import Repay from './components/Repay/Repay';
 import Rates from './components/Rates/Rates';
-import Collateral from './components/Collateral/Collateral';
 import BalanceBorrowUser from './components/BalanceBorrow/BalanceBorrowUser';
 import BalanceBorrowPool from './components/BalanceBorrow/BalanceBorrowPool';
 import Header from './components/Header/Header';
@@ -70,9 +69,6 @@ function App() {
 				</div>
 				<div>
 					<Rates />
-				</div>
-				<div>
-					<Collateral account={accountAddress} />
 				</div>
 				<div>
 					<BalanceBorrowUser account={accountAddress} />
