+++ src/components/ContentUser/Collateral/Collateral.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import { useSubstrate } from '../../substrate-lib';
+import { useSubstrate } from '../../../substrate-lib';
 import { Table, Grid } from 'semantic-ui-react';
-import { UNDERLYING_ASSETS_TYPES } from '../../util/constants';
+import { UNDERLYING_ASSETS_TYPES } from '../../../util/constants';
 import ButtonEnableAndDisable from './ButtonEnableAndDisable';
 
 function Collateral({ account }) {
